refactor: Reorganiza padrões de arquivos
